pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.26.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "streamlit>=1.47.0",
]
